BugFixed:float varlue send by rtos::Queue, modify some wigget for more readability

+++ HMI_FPAC_BACKEND/TouchGFX/gui/include/gui/closedloopgraph_screen/ClosedLoopGraphPresenter.hpp
@@ -35,7 +35,7 @@ public:
     void  sendFeedbackToView();
     void statePidGraphRunP_entry();
     void statePidGraphRunP_exit();
-     void notifyUpdatePidOutput(float pidOutput);
+     void notifyUpdatePidOutput(uint32_t pidOutput);
    
 private:
     ClosedLoopGraphPresenter();
